+++ R/chrome.R
@@ -172,6 +172,13 @@ chrome_print = function(
     if (missing(output) && !file.exists(input))
       output = xfun::with_ext(basename(gsub('[#?].*', '', url)), format)
     output2 = normalizePath(output, mustWork = FALSE)
+    # try to remove the file and throw a clear error if it still exists as it may be locked
+    if (!suppressWarnings(file.remove(output2)) && xfun::file_exists(output2)) {
+      stop(
+        "The file '", output, "' cannot be overwritten",
+        if (format == "pdf") " (may be locked by a PDF reader?)", "."
+      )
+    }
     if (!dir.exists(d <- dirname(output2)) && !dir.create(d, recursive = TRUE)) stop(
       'Cannot create the directory for the output file: ', d
     )